refactor: split rigid and soft body systems
