+++ src/plugin/physics/src/JoltPhysics.hpp
@@ -12,6 +12,8 @@
 #include "system/InitJoltPhysics.hpp"
 #include "system/InitPhysicsManager.hpp"
 #include "system/PhysicsUpdate.hpp"
+#include "system/PhysicsUpdateRigidBody.hpp"
+#include "system/PhysicsUpdateSoftBody.hpp"
 #include "system/ShutdownJoltPhysics.hpp"
 
 #include "utils/BroadPhaseLayerImpl.hpp"
